refactor(adc): return conversion results by value as uint16_t
Replace regular_storage() with regular_result() reading the DMA slot
through a volatile reference and returning the sample by value, narrow
injected_result() to uint16_t, drop the unused injected_storage(), and
enforce uint16_t DMA buffer elements with a static_assert.

+++ apm32/f4/adc/test/adc_test.cpp
@@ -31,7 +31,7 @@ struct adc_traits_1 {
   using stream_type = dma::peripheral_to_memory_stream<
       dma_stream,
       dma_channel,
-      dma::owned_storage<dma::buffer<std::uint32_t, regular_count>>>;
+      dma::owned_storage<dma::buffer<std::uint16_t, regular_count>>>;
   static constexpr nvic::irq_priority dma_irq_priority{4};
   static constexpr auto injected_trigger = inj_trigger{
       .edge = trigger_edge::rising,
